main.dartからMyAppを分離
