+++ lib/ui/home.dart
@@ -9,26 +9,25 @@ import 'dart:convert';
 import 'event_list.dart';
 
 class Home extends StatefulWidget {
-
   @override
   State<StatefulWidget> createState() {
     return _HomeState();
   }
-
 }
 
 class _HomeState extends State<Home> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text("Home"),),
+      appBar: AppBar(
+        title: const Text("Home"),
+      ),
       body: Container(
         child: Column(
           children: [
             FlatButton(
               child: const Text("Twitterでログイン"),
               onPressed: () async {
-
                 final twitterLogin = TwitterLogin(
                   apiKey: "${apiKey}",
                   apiSecretKey: "${apiSecretKey}",
@@ -40,26 +39,24 @@ class _HomeState extends State<Home> {
                   case TwitterLoginStatus.loggedIn:
                     final credential = TwitterAuthProvider.credential(
                         accessToken: authResult.authToken,
-                        secret: authResult.authTokenSecret
-                    );
-                    final firebaseCredential = await FirebaseAuth.instance.signInWithCredential(credential);
+                        secret: authResult.authTokenSecret);
+                    final firebaseCredential = await FirebaseAuth.instance
+                        .signInWithCredential(credential);
 
                     final idToken = await firebaseCredential.user?.getIdToken();
 
                     final request = SessionApiRequest(
-                      token: idToken!,
-                      accessToken: authResult.authToken,
-                      accessTokenSecret: authResult.authTokenSecret
-                    );
+                        token: idToken!,
+                        accessToken: authResult.authToken,
+                        accessTokenSecret: authResult.authTokenSecret);
 
-                    final sessionApiResults = await requestSessionApi(request: request);
+                    final sessionApiResults =
+                        await requestSessionApi(request: request);
                     if (sessionApiResults.status == "OK") {
-                      Navigator.pushReplacement(
-                          context,
+                      Navigator.pushReplacement(context,
                           MaterialPageRoute(builder: (context) {
-                            return EventList();
-                          })
-                      );
+                        return EventList();
+                      }));
                     }
 
                     break;
@@ -68,57 +65,53 @@ class _HomeState extends State<Home> {
                   case TwitterLoginStatus.error:
                     break;
                 }
-
               },
             ),
             RichText(
-                text: TextSpan(
-                    children: [
-                      TextSpan(
-                        text: "アカウントを作成することで",
-                        style: TextStyle(color: Colors.grey[800]),
-                      ),
-                      TextSpan(
-                          text: "利用規約",
-                          style: TextStyle(color: Colors.lightBlue),
-                          recognizer: TapGestureRecognizer()
-                            ..onTap = () {
-                              launch("https://event-follow-front.herokuapp.com/terms");
-                            }
-                      ),
-                      TextSpan(
-                        text: "・",
-                        style: TextStyle(color: Colors.grey[800]),
-                      ),
-                      TextSpan(
-                          text: "プライバシーポリシー",
-                          style: TextStyle(color: Colors.lightBlue),
-                          recognizer: TapGestureRecognizer()
-                            ..onTap = () {
-                              launch("https://event-follow-front.herokuapp.com/privacy_policy");
-                            }
-                      ),
-                      TextSpan(
-                        text: "に同意したものとみなします。",
-                        style: TextStyle(color: Colors.grey[800]),
-                      ),
-                    ]
-                )
-            ),
+                text: TextSpan(children: [
+              TextSpan(
+                text: "アカウントを作成することで",
+                style: TextStyle(color: Colors.grey[800]),
+              ),
+              TextSpan(
+                  text: "利用規約",
+                  style: TextStyle(color: Colors.lightBlue),
+                  recognizer: TapGestureRecognizer()
+                    ..onTap = () {
+                      launch("https://event-follow-front.herokuapp.com/terms");
+                    }),
+              TextSpan(
+                text: "・",
+                style: TextStyle(color: Colors.grey[800]),
+              ),
+              TextSpan(
+                  text: "プライバシーポリシー",
+                  style: TextStyle(color: Colors.lightBlue),
+                  recognizer: TapGestureRecognizer()
+                    ..onTap = () {
+                      launch(
+                          "https://event-follow-front.herokuapp.com/privacy_policy");
+                    }),
+              TextSpan(
+                text: "に同意したものとみなします。",
+                style: TextStyle(color: Colors.grey[800]),
+              ),
+            ])),
           ],
-
         ),
       ),
     );
   }
 }
 
-Future<SessionApiResults> requestSessionApi({ required SessionApiRequest request }) async {
-  final url = Uri.parse("https://event-follow-front.herokuapp.com/api/sessions");
+Future<SessionApiResults> requestSessionApi(
+    {required SessionApiRequest request}) async {
+  final url =
+      Uri.parse("https://event-follow-front.herokuapp.com/api/sessions");
   final response = await http.post(
     url,
     body: json.encode(request.toJson()),
-    headers: { "Content-Type": "application/json" },
+    headers: {"Content-Type": "application/json"},
   );
 
   if (response.statusCode == 200) {
@@ -126,7 +119,6 @@ Future<SessionApiResults> requestSessionApi({ required SessionApiRequest request
   } else {
     throw Exception("Login Failed");
   }
-
 }
 
 class SessionApiRequest {
@@ -141,10 +133,10 @@ class SessionApiRequest {
   });
 
   Map<String, dynamic> toJson() => {
-    "token": this.token,
-    "access_token": this.accessToken,
-    "access_token_secret": this.accessTokenSecret,
-  };
+        "token": this.token,
+        "access_token": this.accessToken,
+        "access_token_secret": this.accessTokenSecret,
+      };
 }
 
 class SessionApiResults {
@@ -157,9 +149,6 @@ class SessionApiResults {
   });
 
   factory SessionApiResults.fromJson(Map<String, dynamic> json) {
-    return SessionApiResults(
-      status: json["status"],
-      message: json["message"]
-    );
+    return SessionApiResults(status: json["status"], message: json["message"]);
   }
 }
